requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
numpy==2.2.1
scikit-learn==1.6.0
pypdf==5.1.0
pytest==8.3.4
